Refactor D3D12 buffer General code clean up

+++ GraphicsEngine/Source/Core/Rendering/Core/VRXEngine.h
@@ -21,7 +21,7 @@ public:
 	static void SetupVRRShader(Shader* S);
 	static void AddVRRToRS(std::vector<ShaderParameter>& S, int lastindex = 0);
 	static void SetupVRSShader(Shader* S);
-	static void AddVRSToRS(std::vector<ShaderParameter>& S, int lastindex = 0);
+	static void AddVRSToRS(std::vector<ShaderParameter>& S, uint64 lastindex = 0);
 private:
 
 };
